This combined commit made several changes to the underlying library including
Converted to use SPIN library

Convert to SPIN library to facilitate the ability to use the T3.x SPI
hardware queue for speed.  This allows me to use any of the SPI busses on
the Teensy.
Teensy 3.2 only has one SPI buss
But T3.5/3.6 both have three queues.

The code uses the hardware SPI queues which allows you to encode CS and DC state
changes into the SPI queue, such that we can run at full speed.  Earlier as with
the ILI9341 display I found using standard pin for CS did not impace performance
very much.  You gained the largest speed by encoding the DC

However since the default configuration for the Teensyview has the DC on
pin 5 which is not a hardware CS pin, I added support for this, while still
using the hardware queue, which appears to be working.  However it is
slightly slower.  For example a S
Also Added support for running on the Teensy LC at a faster speed as
a call to display on hardware DC pin was about 545us and on normal io pin
took about 555us.

Added support for running on T-LC.  Make it sort of emulate the T3.x
code that has queue.  But instead we at least try to full double buffer
the output to keep the queue busy.

Tested on TLC both SPI0 and SPI1
Tested on T3.6 on SPI0, SPI1, SPI2
Earlier tested on T3.2 on SPI0